move block filter out of cc
