+++ pkg/protocol/engine/filter/blockfilter/filter.go
@@ -10,7 +10,7 @@ import (
 	"github.com/iotaledger/iota-core/pkg/model"
 	"github.com/iotaledger/iota-core/pkg/network"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine"
-	"github.com/iotaledger/iota-core/pkg/protocol/engine/congestioncontrol/filter"
+	"github.com/iotaledger/iota-core/pkg/protocol/engine/filter"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/ledger/ledger"
 	iotago "github.com/iotaledger/iota.go/v4"
 )
