+++ pkg/protocol/engine/ledger/ledger/ledger.go
@@ -103,7 +103,7 @@ func New(
 	l := &Ledger{
 		apiProvider:        apiProvider,
 		utxoLedger:         utxoledger.New(utxoStore, apiProvider),
-		accountsLedger:     accountsledger.New(blocksFunc, slotDiffFunc, accountsStore, apiProvider(), protocolParameters.MaxCommitableAge),
+		accountsLedger:     accountsledger.New(blocksFunc, slotDiffFunc, accountsStore, apiProvider(), protocolParameters.MaxCommittableAge),
 		commitmentLoader:   commitmentLoader,
 		conflictDAG:        conflictdagv1.New[iotago.TransactionID, iotago.OutputID, booker.BlockVotePower](committee),
 		errorHandler:       errorHandler,
